+++ src/common/services/recordGenerator/schemas/profiles/monograph/index.ts
@@ -1,7 +1,0 @@
-import { monographInstanceRecordSchema } from './instance.schema';
-import { monographWorkRecordSchema } from './work.schema';
-
-export const monographRecordSchemas = {
-  instance: monographInstanceRecordSchema,
-  work: monographWorkRecordSchema,
-};