+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 * Add support for supplementary content. Refs [UILD-611].
 * Added language subclass support.  Refs [UILD-564].
 * Open Instance using associated profile's Edit/Display form. Refs [UILD-575].
+* Refactor profile handling logic to streamline schema management. Refs [UILD-613].
+* Refactor record generation logic to remove hardcoded profile references. Refs [UILD-614].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -49,6 +51,8 @@
 [UILD-611]:https://folio-org.atlassian.net/browse/UILD-611
 [UILD-564]:https://folio-org.atlassian.net/browse/UILD-564
 [UILD-575]:https://folio-org.atlassian.net/browse/UILD-575
+[UILD-613]:https://folio-org.atlassian.net/browse/UILD-613
+[UILD-614]:https://folio-org.atlassian.net/browse/UILD-614
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].